# src/main/java/com/cresloga/inventory/household/householdinventory/models/dto/Items.java
package com.cresloga.inventory.household.householdinventory.models.dto;

import java.util.Objects;
import java.io.Serializable;
import java.util.ArrayList;
import org.springframework.validation.annotation.Validated;

/**
 * Items
 */
@Validated
@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2020-05-02T11:32:34.017Z[GMT]")
public class Items extends ArrayList<Item> implements Serializable {

	/**
	 * 
	 */
	private static final long serialVersionUID = -142336116948723556L;

	@Override
	public boolean equals(java.lang.Object o) {
		if (this == o) {
			return true;
		}
		if (o == null || getClass() != o.getClass()) {
			return false;
		}
		return true;
	}

	@Override
	public int hashCode() {
		return Objects.hash(super.hashCode());
	}

	@Override
	public String toString() {
		StringBuilder sb = new StringBuilder();
		sb.append("class Items {\n");
		sb.append("    ").append(toIndentedString(super.toString())).append("\n");
		sb.append("}");
		return sb.toString();
	}

	/**
	 * Convert the given object to string with each line indented by 4 spaces
	 * (except the first line).
	 */
	private String toIndentedString(java.lang.Object o) {
		if (o == null) {
			return "null";
		}
		return o.toString().replace("\n", "\n    ");
	}
}